refactor: mudando o grafico saldo para a pasta ui componentes
