+++ app/src/main/resources/perfil_cliente.fxml
@@ -5,6 +5,7 @@
 <?import javafx.scene.layout.*?>
 <?import planejadorfinanceiro.model.*?>
 
+<?import planejadorfinanceiro.ui.componentes.GraficoSaldo?>
 <VBox alignment="CENTER" prefHeight="681.0" prefWidth="700.0" spacing="15" style="-fx-padding: 20;" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="planejadorfinanceiro.PerfilClienteController">
     <Label style="-fx-font-size: 18; -fx-font-weight: bold;" text="Perfil do Cliente" />
